requirements.txt
requests
python-dotenv 
google-generativeai 
colorama 
slither-analyzer 
tenacity  
pytest 
web3 
reportlab